+++ propulsion-consumer/Program.fs
@@ -49,7 +49,7 @@ module CmdParser =
                 | MaxInflightMb _ ->        "maximum MiB of data to read ahead. Default: 10."
                 | LagFreqM _ ->             "specify frequency (minutes) to dump lag stats. Default: off."
 
-                | MaxDop _ ->               "maximum number of items to process in parallel. Default: 1024."
+                | MaxDop _ ->               "maximum number of items to process in parallel. Default: 8"
                 | Verbose _ ->              "request verbose logging."
     type Arguments(a : ParseResults<Parameters>) =
         member __.Broker =                  a.TryGetResult Broker |> defaultWithEnvVar "PROPULSION_KAFKA_BROKER" "Broker" |> Uri
@@ -58,7 +58,7 @@ module CmdParser =
         member __.MaxInFlightBytes =        a.GetResult(MaxInflightMb, 10.) * 1024. * 1024. |> int64
         member __.LagFrequency =            a.TryGetResult LagFreqM |> Option.map TimeSpan.FromMinutes
 
-        member __.MaxDop =                  match a.TryGetResult MaxDop with Some x -> x | None -> 1024
+        member __.MaxDop =                  a.GetResult(MaxDop, 8)
         member __.Verbose =                 a.Contains Verbose
 
     /// Parse the commandline; can throw exceptions in response to missing arguments and/or `-h`/`--help` args